add: start implementing TCP teardown
Not sure how to know that everything got sent.

For instance, how do we know that the server got the initial FIN? We
depend on the FINACK.
How do we know the future FINACK will be received? We depend on a future ACK.
How do we know this future ACK will be received? We depend on... ?
So on and so forth.

This seems to be an endless loop.